fix: type corrections and housekeeping
- compliance.go + models/compliance.go: TLSIssues int → uint64 to
  match ClickHouse count() return type and avoid scan mismatch
- models/agent.go: FlowCount1h int64 → uint64 (counts are non-negative)
- agent/Cargo.lock: update auto-generated dependency lock file
- desktop/.claude/launch.json: rename debug config to "desktop-dev"

+++ hub/api/models/compliance.go
@@ -51,7 +51,7 @@ type TopTalker struct {
 type ComplianceSummary struct {
 	TotalConnections    uint64 `json:"total_connections"`
 	ExternalConnections uint64 `json:"external_connections"`
-	TLSIssues           int    `json:"tls_issues"`
+	TLSIssues           uint64 `json:"tls_issues"`
 	TopTalkerCount      int    `json:"top_talker_count"`
 	Window              string `json:"window"`
 }
